+++ src/components/manage/Blocks/Listing/SummaryTemplate.jsx
@@ -5,6 +5,7 @@ import { flattenToAppURL } from '@plone/volto/helpers';
 
 import DefaultImageSVG from '@plone/volto/components/manage/Blocks/Listing/default-image.svg';
 import { isInternalURL } from '@plone/volto/helpers/Url/Url';
+import Image from '@plone/volto/components/theme/Image/Image';
 
 const SummaryTemplate = ({ items, linkTitle, linkHref, isEditMode }) => {
   let link = null;
@@ -26,13 +27,20 @@ const SummaryTemplate = ({ items, linkTitle, linkHref, isEditMode }) => {
         {items.map((item) => (
           <div className="listing-item" key={item['@id']}>
             <ConditionalLink item={item} condition={!isEditMode}>
-              {!item.image_field && <img src={DefaultImageSVG} alt="" />}
+              {!item.image_field && (
+                <picture>
+                  <img
+                    className="volto-image thumb"
+                    src={DefaultImageSVG}
+                    alt=""
+                  />
+                </picture>
+              )}
               {item.image_field && (
-                <img
-                  src={flattenToAppURL(
-                    `${item['@id']}/@@images/${item.image_field}/preview`,
-                  )}
-                  alt={item.title}
+                <Image
+                  image={item['@id']}
+                  size="thumb"
+                  imageField={item.image_field}
                 />
               )}
               <div className="listing-body">