+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   a `blockSchema` with a schema factory. This schema will also be used to
   extract a default block value.
 
+- Add Image with srcset and lazy loading using Plone scales @nzambello
 - Add runtime configuration for `@babel/plugin-transform-react-jsx` set to `automatic`. This enables the new JSX runtime: https://reactjs.org/blog/2020/09/22/introducing-the-new-jsx-transform.html So no longer `import React from 'react'` is needed anymore.
 - Update favicon and related tags with best practices @sneridagh